Fixed bug with multiple balls
When one of the multiple balls hits the powerup for new balls, the engine didn't know which ball that was.
Also, sometimes the ball variable wouldn't hold the last remaining ball from many balls which led to a bug when calling `ball.reset()`.

Now, there is no "global" ball variable.